Moved types to Domain/Services/Models Renamed EpicVisualBlock to EpicCard
