--- src/JiraEpicRoadmapper.UI/Models/IEpicCardPainter.cs
@@ -1,7 +1,0 @@
-﻿namespace JiraEpicRoadmapper.UI.Models
-{
-    public interface IEpicCardPainter
-    {
-        string GetColor(EpicVisualBlock block);
-    }
-}
