--- src/JiraEpicRoadmapper.UI/Models/IStatusVisualizer.cs
@@ -1,9 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace JiraEpicRoadmapper.UI.Models
-{
-    public interface IStatusVisualizer
-    {
-        IEnumerable<string> GetStatusIcons(EpicVisualBlock block);
-    }
-}
