+++ src/JiraEpicRoadmapper.UI/Pages/RoadmapView.razor
@@ -26,7 +26,7 @@
     private Epic[] _epics;
     private string _errors;
     //TODO: test
-    public EpicVisualBlock SelectedBlock { get; private set; }
+    public EpicCard SelectedBlock { get; private set; }
 
 
     protected override async Task OnInitializedAsync()
@@ -58,7 +58,7 @@
         await JsRuntime.InvokeVoidAsync("scroll", CancellationToken.None, xPosition - LayoutSettings.DaySpan, 0);
     }
 
-    private void HandleCardSelect(EpicVisualBlock card)
+    private void HandleCardSelect(EpicCard card)
     {
         SelectedBlock = card;
     }
