+++ src/JiraEpicRoadmapper.UI/Services/IEpicCardPainter.cs
@@ -1,0 +1,9 @@
+﻿using JiraEpicRoadmapper.UI.Models;
+
+namespace JiraEpicRoadmapper.UI.Services
+{
+    public interface IEpicCardPainter
+    {
+        string GetColor(EpicCard block);
+    }
+}
