+++ src/JiraEpicRoadmapper.UI/Services/ILayoutDesigner.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
+using JiraEpicRoadmapper.UI.Models;
 
-namespace JiraEpicRoadmapper.UI.Models
+namespace JiraEpicRoadmapper.UI.Services
 {
     public interface ILayoutDesigner
     {
