+++ src/JiraEpicRoadmapper.UI/Services/IStatusVisualizer.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+using JiraEpicRoadmapper.UI.Models;
+
+namespace JiraEpicRoadmapper.UI.Services
+{
+    public interface IStatusVisualizer
+    {
+        IEnumerable<string> GetStatusIcons(EpicCard block);
+    }
+}
