+++ src/JiraEpicRoadmapper.UI/Shared/ControlPanel.razor
@@ -14,12 +14,12 @@
 
 @code {
     //TODO: test the component!
-    private EpicVisualBlock _selectedBlock;
+    private EpicCard _selectedBlock;
     private bool PanelVisible { get; set; }
 
 
     [Parameter]
-    public EpicVisualBlock SelectedBlock
+    public EpicCard SelectedBlock
     {
         get => _selectedBlock;
         set
