+++ test/JiraEpicRoadmapper.UI.UnitTests/ProjectLayoutTests.cs
@@ -1,7 +1,9 @@
 ﻿using System;
 using System.Linq;
 using JiraEpicRoadmapper.Contracts;
+using JiraEpicRoadmapper.UI.Domain;
 using JiraEpicRoadmapper.UI.Models;
+using JiraEpicRoadmapper.UI.Services;
 using Moq;
 using Shouldly;
 using Xunit;
